third_party/blink/web_tests/svg/dynamic-updates/SVGFECompositeElement-dom-k3-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'k3' attribute of the SVGFECompositeElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS overComposite1.getAttribute('k3') is "2.9"
PASS overComposite1.getAttribute('k3') is ".1"
PASS successfullyParsed is true

TEST COMPLETE
